backend/requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
django-cors-headers==4.3.1

# Environment
python-dotenv==1.0.1

# Geospatial
django.contrib.gis

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2

# Caching
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.1

# Development
django-debug-toolbar==4.3.0
